fix up frontend to work with refactored backend, add useSearch composable and SearchResults component
